+++ lib/sequel-rails/storage/mysql2.rb
@@ -1,0 +1,8 @@
+module Rails
+  module Sequel
+    module Storage
+      class Mysql2 < Mysql
+      end
+    end
+  end
+end